move all witneess components to one file
