+++ src/web/components/elements/WitnessBlockComponent.stories.tsx
@@ -3,12 +3,32 @@ import { css } from 'emotion';
 
 import { Pillar } from '@guardian/types';
 
-import { WitnessImageBlockComponent } from './WitnessImageBlockComponent';
+import {
+	WitnessTextBlockComponent,
+	WitnessImageBlockComponent,
+	WitnessVideoBlockComponent,
+} from './WitnessBlockComponent';
 
 export default {
-	component: WitnessImageBlockComponent,
-	title: 'Components/WitnessImageBlockComponent',
+	title: 'Components/WitnessBlockComponent',
 };
+
+export const WitnessTextBlockComponentDefault = () => (
+	<div
+		className={css`
+			margin: 15px;
+			width: 620px;
+		`}
+	>
+		<WitnessTextBlockComponent
+			title="Back in the 1970&#39;s"
+			description='You could dial first and then push in ur 2p. I never had any change and needed a lift. My parents knew the routine, I&#39;d dial and you could get a second or two to shout down the line "Pick me up" Ah how I miss the days of a simple life where no one knew where I was and I could get away with anything as long as I was in the door by 10pm! '
+			authorName="Louise Griffiths"
+			dateCreated="2016-01-29T22:19:51Z"
+			pillar={Pillar.Sport}
+		/>
+	</div>
+);
 
 const assets: WitnessAssetType[] = [
 	{
@@ -99,6 +119,24 @@ export const WitnessImageBlockComponentDefault = () => (
 			authorName="Nick Ellis"
 			dateCreated="2015-08-25T12:20:58Z"
 			alt="Risk - Home Made Space expansion"
+			pillar={Pillar.Sport}
+		/>
+	</div>
+);
+
+export const WitnessVideoBlockComponentDefault = () => (
+	<div
+		className={css`
+			margin: 15px;
+			width: 620px;
+		`}
+	>
+		<WitnessVideoBlockComponent
+			title="Warhammer Quest Timelapse"
+			description="Tell me if you&#39;ve heard this one before: A Dwarf, a thief, a fighter and a priest walk into a dungeon ... Warhammer Quest is the quintessential &#39;Old School&#39; dungeon-crawler."
+			authorName="Gregg Lewis-Qualls"
+			youtubeHtml='<iframe width="440" height="330" src="https://www.youtube.com/embed/XRq83txBgxw?wmode=opaque&feature=oembed" frameborder="0" allowfullscreen></iframe>'
+			dateCreated="2015-08-27T13:32:32Z"
 			pillar={Pillar.Sport}
 		/>
 	</div>
